moving Components to own folder
